fix code duplications and code smells from sonarqube report
