+++ package.json
@@ -4,7 +4,7 @@
   "license": "GPL-3.0-only",
   "private": true,
   "dependencies": {
-    "@drill4j/socket": "0.1.0",
+    "@drill4j/socket": "0.1.2",
     "@redneckz/react-bem-helper": "2.0.3",
     "@types/jest": "24.0.11",
     "@types/node": "11.12.2",
