+++ src/common/connection/default-ws-connection.ts
@@ -1,9 +1,15 @@
 import { DrillSocket } from '@drill4j/socket';
 
-// process.env.REACT_APP_ENV
-//   ? `wss://${window.location.host}/ws/${socket}`
-//   : `wss://localhost:8443/ws/${socket}`,
+import { TOKEN_KEY } from '../constants';
 
-export const defaultAdminSocket = new DrillSocket();
+const getSocketUrl = (socket: string) => {
+  const token = localStorage.getItem(TOKEN_KEY);
 
-export const defaultPluginSocket = new DrillSocket('drill-plugin-socket');
+  return `wss://${
+    process.env.REACT_APP_ENV ? window.location.host : 'localhost:8443'
+  }/ws/${socket}?token=${token}`;
+};
+
+export const defaultAdminSocket = new DrillSocket(getSocketUrl('drill-admin-socket'));
+
+export const defaultPluginSocket = new DrillSocket(getSocketUrl('drill-plugin-socket'));
