+++ src/forms/inputs/dropdown-input/dropdown-input.tsx
@@ -2,6 +2,7 @@ import * as React from 'react';
 import { BEM, div } from '@redneckz/react-bem-helper';
 
 import { Icons } from '../../../components';
+import { useClickOutside } from '../../../hooks';
 import { DropdownItem } from './dropdown-item';
 
 import styles from './dropdown-input.module.scss';
@@ -19,19 +20,7 @@ const dropdownInput = BEM(styles);
 export const DropdownInput = dropdownInput(
   ({ className, options, value, onChange, error }: Props) => {
     const [isExpanded, setIsExpanded] = React.useState(false);
-    const node = React.useRef<HTMLDivElement>(null);
-    React.useEffect(() => {
-      function handleClick(event: any) {
-        if (node && node.current && node.current.contains(event.target)) {
-          return;
-        }
-        setIsExpanded(false);
-      }
-      document.addEventListener('mousedown', handleClick);
-      return () => {
-        document.removeEventListener('mousedown', handleClick);
-      };
-    }, []);
+    const node = useClickOutside(() => setIsExpanded(false));
     return (
       <div className={className} ref={node}>
         <Container onClick={() => setIsExpanded(!isExpanded)}>
