+++ src/hooks/index.ts
@@ -1,1 +1,2 @@
 export { useWsConnection } from './use-ws-connection';
+export { useClickOutside } from './use-click-outside';
