+++ src/hooks/use-click-outside.ts
@@ -1,0 +1,20 @@
+import { useRef, useEffect } from 'react';
+
+export function useClickOutside(callback: () => void) {
+  const node = useRef<HTMLDivElement>(null);
+
+  useEffect(() => {
+    function handleClick(event: any) {
+      if (node && node.current && node.current.contains(event.target)) {
+        return;
+      }
+      callback();
+    }
+    document.addEventListener('mousedown', handleClick);
+    return () => {
+      document.removeEventListener('mousedown', handleClick);
+    };
+  }, []);
+
+  return node;
+}
